move test_features
