+++ tests/bofire/data_models/domain/test_inputs.py
@@ -1000,3 +1000,103 @@ def test_inputs_get_bounds_fit():
     assert opt_bounds[1][-2] == 0
     assert fit_bounds[1][-1] == 1
     assert fit_bounds[1][-2] == 1
+
+
+@pytest.mark.skipif(not RDKIT_AVAILABLE, reason="requires rdkit")
+@pytest.mark.parametrize(
+    "specs, molecular_keys, continuous_keys, expected_molecular_indices, expected_continuous_indices",
+    [
+        (
+            {
+                "x2": CategoricalEncodingEnum.ONE_HOT,
+                "x3": CategoricalEncodingEnum.ONE_HOT,
+                "x4": Fingerprints(n_bits=2),
+            },
+            ["x4"],
+            ["x1"],
+            [0, 1],
+            [2],
+        ),
+        (
+            {
+                "x2": CategoricalEncodingEnum.ONE_HOT,
+                "x3": CategoricalEncodingEnum.ONE_HOT,
+                "x4": Fragments(fragments=["fr_unbrch_alkane", "fr_thiocyan"]),
+            },
+            ["x4"],
+            ["x1"],
+            [0, 1],
+            [2],
+        ),
+        (
+            {
+                "x2": CategoricalEncodingEnum.ONE_HOT,
+                "x3": CategoricalEncodingEnum.ONE_HOT,
+                "x4": MordredDescriptors(descriptors=["NssCH2", "ATSC2d"]),
+            },
+            [],
+            ["x4", "x1"],
+            [],
+            [0, 1, 2],
+        ),
+        (
+            {
+                "x2": CategoricalEncodingEnum.ONE_HOT,
+                "x3": CategoricalEncodingEnum.DESCRIPTOR,
+                "x4": Fingerprints(n_bits=2),
+            },
+            ["x4"],
+            ["x1", "x3"],
+            [0, 1],
+            [2, 3, 4],
+        ),
+        (
+            {
+                "x2": CategoricalEncodingEnum.ONE_HOT,
+                "x3": CategoricalEncodingEnum.DESCRIPTOR,
+                "x4": Fragments(fragments=["fr_unbrch_alkane", "fr_thiocyan"]),
+            },
+            ["x4"],
+            ["x1", "x3"],
+            [0, 1],
+            [2, 3, 4],
+        ),
+        (
+            {
+                "x2": CategoricalEncodingEnum.ONE_HOT,
+                "x3": CategoricalEncodingEnum.DESCRIPTOR,
+                "x4": MordredDescriptors(descriptors=["NssCH2", "ATSC2d"]),
+            },
+            [],
+            ["x4", "x1", "x3"],
+            [],
+            [0, 1, 2, 3, 4],
+        ),
+    ],
+)
+def test_inputs_get_feature_indices(
+    specs,
+    molecular_keys,
+    continuous_keys,
+    expected_molecular_indices,
+    expected_continuous_indices,
+):
+    inps = Inputs(
+        features=[
+            ContinuousInput(key="x1", bounds=(0, 1)),
+            CategoricalInput(key="x2", categories=["apple", "banana", "orange"]),
+            CategoricalDescriptorInput(
+                key="x3",
+                categories=["apple", "banana", "orange", "cherry"],
+                descriptors=["d1", "d2"],
+                values=[[1, 2], [3, 4], [5, 6], [7, 8]],
+            ),
+            MolecularInput(key="x4"),
+        ]
+    )
+
+    mol_dims = inps.get_feature_indices(specs, molecular_keys)
+    ord_dims = inps.get_feature_indices(specs, continuous_keys)
+
+    assert mol_dims == expected_molecular_indices
+    assert ord_dims == expected_continuous_indices
